Moved get_neighbors into actions.py.
