+++ capital/bigbox.py
@@ -5,9 +5,9 @@ will drive small retailers out of business.
 
 import random
 
+import lib.actions as acts
 from lib.agent import MOVE, Agent, join
 from lib.display_methods import BLACK, BLUE, GREEN, RED, ORANGE, PURPLE
-import lib.model as mdl
 from lib.model import Model
 from lib.model import NUM_MBRS, MBR_ACTION, COLOR, MBR_CREATOR
 import registry.registry as reg
@@ -139,7 +139,7 @@ def consumer_action(consumer, **kwargs):
     item_needed = consumer.get_attr(ITEM_NEEDED)
     box = reg.get_model(consumer.exec_key)
     hood_size = box.get_prop("hood_size", DEF_HOOD_SIZE)
-    sellers = mdl.get_neighbors(consumer, pred=sells_good, size=hood_size)
+    sellers = acts.get_neighbors(consumer, pred=sells_good, size=hood_size)
     shop_at = choose_store(consumer, sellers.members.items())
     if shop_at is not None:
         transaction(shop_at, consumer)
