+++ models/act_in.py
@@ -2,10 +2,10 @@
 This model tries to demonstrate how patterns forms in a closed system
 """
 
+import lib.actions as acts
 from lib.agent import MOVE, Agent
 from lib.display_methods import RED, BLUE
 from lib.model import Model, NUM_MBRS, MBR_ACTION, NUM_MBRS_PROP, COLOR
-from lib.space import get_neighbors
 from lib.utils import Debug
 from registry.registry import save_reg
 
@@ -35,8 +35,7 @@ def agent_action(agent, **kwargs):
     if DEBUG.debug:
         print("Agent {} is located at {}".format(agent.name,
                                                  agent.get_pos()))
-    neighbors = get_neighbors(agent)
-    for neighbor in neighbors:
+    for neighbor in acts.get_neighbors(agent):
         print(f"{str(agent)} has neighbor {str(neighbor)}")
     return MOVE
